use a signed int for cheat compare value
-1 means no compare value
0-0xFF is the valid range of compare value

+++ Core/CheatManager.h
@@ -26,7 +26,7 @@ struct CheatInfo
 	uint32_t Address;
 	char GameGenieCode[9];
 	uint8_t Value;
-	uint8_t CompareValue;
+	int16_t CompareValue;
 	bool UseCompareValue;
 	bool IsRelativeAddress;
 };
